# openapi.yaml
openapi: 3.1.0
info:
  title: Chronometrics API (contract draft)
  version: 0.1.0
  description: Root OpenAPI spec for contract work (draft). Contains events, bulk publish, metrics, and health endpoints.
servers:
  - url: https://api.example.com
paths:
  /events:
    post:
      summary: Publish a single event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventPayload'
      responses:
        '202':
          description: Accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Retry-After:
              description: Seconds until client should retry
              schema:
                type: integer
  /events/bulk:
    post:
      summary: Publish multiple events in a single request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRequest'
      responses:
        '202':
          description: Accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Retry-After:
              description: Seconds until client should retry
              schema:
                type: integer
  /metrics:
    get:
      summary: Query aggregated metrics
      parameters:
        - name: event_name
          in: query
          required: true
          description: Event name to query
          schema:
            type: string
        - name: from
          in: query
          required: true
          description: Start timestamp (epoch ms or epoch seconds)
          schema:
            type: integer
        - name: to
          in: query
          required: true
          description: End timestamp (epoch ms or epoch seconds)
          schema:
            type: integer
        - name: group_by
          in: query
          description: Dimension to group results by
          schema:
            type: string
            enum: [channel, hour, day]
      responses:
        '200':
          description: Metrics result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricResult'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /health:
    get:
      summary: Service health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
components:
  schemas:
    EventPayload:
      type: object
      required:
        - event_name
        - user_id
        - timestamp
      properties:
        event_name:
          type: string
        user_id:
          type: string
        timestamp:
          type: integer
          description: epoch milliseconds or epoch seconds
        channel:
          type: string
        campaign_id:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
    BulkRequest:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventPayload'
    BulkResponse:
      type: object
      properties:
        accepted:
          type: integer
        rejected:
          type: integer
        errors:
          type: array
          items:
            $ref: '#/components/schemas/BulkItemError'
    BulkItemError:
      type: object
      properties:
        index:
          type: integer
        error:
          type: string
        code:
          type: string
    MetricResult:
      type: object
      properties:
        event_name:
          type: string
        from:
          type: integer
        to:
          type: integer
        group_by:
          type: string
        total_count:
          type: integer
        unique_count:
          type: integer
        groups:
          type: array
          items:
            $ref: '#/components/schemas/GroupResult'
    GroupResult:
      type: object
      properties:
        key:
          type: string
        total_count:
          type: integer
        unique_count:
          type: integer
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, down]
        buffer_strategy:
          type: string
        uptime_seconds:
          type: integer
        clickhouse:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, PUBLISH_FAILED, PAYLOAD_TOO_LARGE, QUERY_TIMEOUT, INTERNAL_ERROR, READ_ERROR, METHOD_NOT_ALLOWED, UNSUPPORTED_MEDIA_TYPE]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    AcceptedResponse:
      type: object
      properties:
        status:
          type: string
          example: accepted